+++ src/SlobSteak.Application/ServiceCollectionExtensions.cs
@@ -13,6 +13,7 @@ public static class ServiceCollectionExtensions
     {
         services.AddScoped<SeedAdminService>();
         services.AddScoped<LoginService>();
+        services.AddScoped<ChangePasswordService>();
 
         return services;
     }